+++ packages/app_open_example/lib/app_lifecycle_reactor.dart
@@ -14,18 +14,23 @@
 
 // ignore_for_file: public_member_api_docs
 import 'package:app_open_example/app_open_ad_manager.dart';
-import 'package:flutter/material.dart';
+import 'package:google_mobile_ads/google_mobile_ads.dart';
 
-class AppLifecycleReactor extends WidgetsBindingObserver {
+/// Listens for app foreground events and shows app open ads.
+class AppLifecycleReactor {
   final AppOpenAdManager appOpenAdManager;
 
   AppLifecycleReactor({required this.appOpenAdManager});
 
-  @override
-  Future<void> didChangeAppLifecycleState(AppLifecycleState state) async {
-    // Try to show an app open ad if the app is being resumed and
-    // we're not already showing an app open ad.
-    if (state == AppLifecycleState.resumed) {
+  void listenToAppStateChanges() {
+    AppStateEventNotifier.startListening();
+    AppStateEventNotifier.appStateStream
+        .forEach((state) => _onAppStateChanged(state));
+  }
+
+  void _onAppStateChanged(AppState appState) {
+    print('New AppState state: $appState');
+    if (appState == AppState.foreground) {
       appOpenAdManager.showAdIfAvailable();
     }
   }