+++ packages/google_mobile_ads/android/src/main/java/io/flutter/plugins/googlemobileads/FlutterInterstitialAd.java
@@ -48,7 +48,7 @@ class FlutterInterstitialAd extends FlutterAd.FlutterOverlayAd {
   void load() {
     if (manager != null && adUnitId != null && request != null) {
       flutterAdLoader.loadInterstitial(
-          adUnitId, request.asAdRequest(), new DelegatingInterstitialAdLoadCallback(this));
+          adUnitId, request.asAdRequest(adUnitId), new DelegatingInterstitialAdLoadCallback(this));
     }
   }