+++ packages/google_mobile_ads/android/src/main/java/io/flutter/plugins/googlemobileads/GoogleMobileAdsPlugin.java
@@ -59,9 +59,10 @@ public class GoogleMobileAdsPlugin implements FlutterPlugin, ActivityAware, Meth
   // This is always null when not using v2 embedding.
   @Nullable private FlutterPluginBinding pluginBinding;
   @Nullable private AdInstanceManager instanceManager;
-  @Nullable private ActivityPluginBinding activityBinding;
   @Nullable private AdMessageCodec adMessageCodec;
+  @Nullable private AppStateNotifier appStateNotifier;
   private final Map<String, NativeAdFactory> nativeAdFactories = new HashMap<>();
+  @Nullable private MediationNetworkExtrasProvider mediationNetworkExtrasProvider;
   private final FlutterMobileAdsWrapper flutterMobileAds;
   /**
    * Public constructor for the plugin. Dependency initialization is handled in lifecycle methods
@@ -82,6 +83,12 @@ public class GoogleMobileAdsPlugin implements FlutterPlugin, ActivityAware, Meth
     this.flutterMobileAds = flutterMobileAds;
   }
 
+  @VisibleForTesting
+  protected GoogleMobileAdsPlugin(@NonNull AppStateNotifier appStateNotifier) {
+    this.appStateNotifier = appStateNotifier;
+    this.flutterMobileAds = null;
+  }
+
   /**
    * Interface used to display a {@link com.google.android.gms.ads.nativead.NativeAd}.
    *
@@ -98,7 +105,7 @@ public class GoogleMobileAdsPlugin implements FlutterPlugin, ActivityAware, Meth
      * @param customOptions Used to pass additional custom options to create the {@link
      *     com.google.android.gms.ads.nativead.NativeAdView}. Nullable.
      * @return a {@link com.google.android.gms.ads.nativead.NativeAdView} that is overlaid on top of
-     *     the FlutterView.
+     *     the FlutterView
      */
     NativeAdView createNativeAd(NativeAd nativeAd, Map<String, Object> customOptions);
   }
@@ -108,18 +115,61 @@ public class GoogleMobileAdsPlugin implements FlutterPlugin, ActivityAware, Meth
    * used to create {@link com.google.android.gms.ads.nativead.NativeAdView}s from a Native Ad
    * created in Dart.
    *
-   * @param engine maintains access to a GoogleMobileAdsPlugin instance.
+   * @param engine maintains access to a GoogleMobileAdsPlugin instance
    * @param factoryId a unique identifier for the ad factory. The Native Ad created in Dart includes
    *     a parameter that refers to this.
    * @param nativeAdFactory creates {@link com.google.android.gms.ads.nativead.NativeAdView}s when
-   *     Flutter NativeAds are created.
-   * @return whether the factoryId is unique and the nativeAdFactory was successfully added.
+   *     Flutter NativeAds are created
+   * @return whether the factoryId is unique and the nativeAdFactory was successfully added
    */
   public static boolean registerNativeAdFactory(
       FlutterEngine engine, String factoryId, NativeAdFactory nativeAdFactory) {
     final GoogleMobileAdsPlugin gmaPlugin =
         (GoogleMobileAdsPlugin) engine.getPlugins().get(GoogleMobileAdsPlugin.class);
     return registerNativeAdFactory(gmaPlugin, factoryId, nativeAdFactory);
+  }
+
+  /**
+   * Registers a {@link MediationNetworkExtrasProvider} used to provide network extras to the plugin
+   * when it creates ad requests.
+   *
+   * @param engine The {@link FlutterEngine} which should have an attached instance of this plugin
+   * @param mediationNetworkExtrasProvider The {@link MediationNetworkExtrasProvider} which will be
+   *     used to provide network extras when ad requests are created
+   * @return whether {@code mediationNetworkExtrasProvider} was registered to a {@code
+   *     GoogleMobileAdsPlugin} associated with {@code engine}
+   */
+  public static boolean registerMediationNetworkExtrasProvider(
+      FlutterEngine engine, MediationNetworkExtrasProvider mediationNetworkExtrasProvider) {
+    final GoogleMobileAdsPlugin gmaPlugin =
+        (GoogleMobileAdsPlugin) engine.getPlugins().get(GoogleMobileAdsPlugin.class);
+    if (gmaPlugin == null) {
+      return false;
+    }
+    gmaPlugin.mediationNetworkExtrasProvider = mediationNetworkExtrasProvider;
+    if (gmaPlugin.adMessageCodec != null) {
+      gmaPlugin.adMessageCodec.setMediationNetworkExtrasProvider(mediationNetworkExtrasProvider);
+    }
+    return true;
+  }
+
+  /**
+   * Unregisters any {@link MediationNetworkExtrasProvider} that have been previously registered
+   * with the plugin using {@code unregisterMediationNetworkExtrasProvider}.
+   *
+   * @param engine The {@link FlutterEngine} which should have an attached instance of this plugin
+   */
+  public static void unregisterMediationNetworkExtrasProvider(FlutterEngine engine) {
+    final GoogleMobileAdsPlugin gmaPlugin =
+        (GoogleMobileAdsPlugin) engine.getPlugins().get(GoogleMobileAdsPlugin.class);
+    if (gmaPlugin == null) {
+      return;
+    }
+
+    if (gmaPlugin.adMessageCodec != null) {
+      gmaPlugin.adMessageCodec.setMediationNetworkExtrasProvider(null);
+    }
+    gmaPlugin.mediationNetworkExtrasProvider = null;
   }
 
   private static boolean registerNativeAdFactory(
@@ -140,12 +190,12 @@ public class GoogleMobileAdsPlugin implements FlutterPlugin, ActivityAware, Meth
    * used to create {@link com.google.android.gms.ads.nativead.NativeAdView}s from a Native Ad
    * created in Dart.
    *
-   * @param engine maintains access to a GoogleMobileAdsPlugin instance.
+   * @param engine maintains access to a GoogleMobileAdsPlugin instance
    * @param factoryId a unique identifier for the ad factory. The Native ad created in Dart includes
-   *     a parameter that refers to this.
+   *     a parameter that refers to this
    * @return the previous {@link
    *     io.flutter.plugins.googlemobileads.GoogleMobileAdsPlugin.NativeAdFactory} associated with
-   *     this factoryId, or null if there was none for this factoryId.
+   *     this factoryId, or null if there was none for this factoryId
    */
   public static NativeAdFactory unregisterNativeAdFactory(FlutterEngine engine, String factoryId) {
     final FlutterPlugin gmaPlugin = engine.getPlugins().get(GoogleMobileAdsPlugin.class);
@@ -177,6 +227,9 @@ public class GoogleMobileAdsPlugin implements FlutterPlugin, ActivityAware, Meth
   public void onAttachedToEngine(FlutterPluginBinding binding) {
     pluginBinding = binding;
     adMessageCodec = new AdMessageCodec(binding.getApplicationContext());
+    if (mediationNetworkExtrasProvider != null) {
+      adMessageCodec.setMediationNetworkExtrasProvider(mediationNetworkExtrasProvider);
+    }
     final MethodChannel channel =
         new MethodChannel(
             binding.getBinaryMessenger(),
@@ -189,16 +242,19 @@ public class GoogleMobileAdsPlugin implements FlutterPlugin, ActivityAware, Meth
         .registerViewFactory(
             "plugins.flutter.io/google_mobile_ads/ad_widget",
             new GoogleMobileAdsViewFactory(instanceManager));
+    appStateNotifier = new AppStateNotifier(binding.getBinaryMessenger());
   }
 
   @Override
   public void onDetachedFromEngine(FlutterPluginBinding binding) {
-    // Do nothing
+    if (appStateNotifier != null) {
+      appStateNotifier.stop();
+      appStateNotifier = null;
+    }
   }
 
   @Override
   public void onAttachedToActivity(ActivityPluginBinding binding) {
-    activityBinding = binding;
     if (instanceManager != null) {
       instanceManager.setActivity(binding.getActivity());
     }
@@ -244,7 +300,11 @@ public class GoogleMobileAdsPlugin implements FlutterPlugin, ActivityAware, Meth
       Log.e(TAG, "method call received before instanceManager initialized: " + call.method);
       return;
     }
-    Context appContext = pluginBinding.getApplicationContext();
+    // Use activity as context if available.
+    Context context =
+        (instanceManager.getActivity() != null)
+            ? instanceManager.getActivity()
+            : pluginBinding.getApplicationContext();
     switch (call.method) {
       case "_init":
         // Internal init. This is necessary to cleanup state on hot restart.
@@ -253,7 +313,10 @@ public class GoogleMobileAdsPlugin implements FlutterPlugin, ActivityAware, Meth
         break;
 
       case "MobileAds#initialize":
-        flutterMobileAds.initialize(appContext, new FlutterInitializationListener(result));
+        flutterMobileAds.initialize(context, new FlutterInitializationListener(result));
+        break;
+      case "MobileAds#getRequestConfiguration":
+        result.success(flutterMobileAds.getRequestConfiguration());
         break;
       case "MobileAds#updateRequestConfiguration":
         RequestConfiguration.Builder builder = MobileAds.getRequestConfiguration().toBuilder();
@@ -284,7 +347,7 @@ public class GoogleMobileAdsPlugin implements FlutterPlugin, ActivityAware, Meth
                 call.<String>argument("adUnitId"),
                 call.<FlutterAdRequest>argument("request"),
                 call.<FlutterAdSize>argument("size"),
-                getBannerAdCreator(appContext));
+                getBannerAdCreator(context));
         instanceManager.trackAd(bannerAd, call.<Integer>argument("adId"));
         bannerAd.load();
         result.success(null);
@@ -308,7 +371,7 @@ public class GoogleMobileAdsPlugin implements FlutterPlugin, ActivityAware, Meth
                 .setCustomOptions(call.<Map<String, Object>>argument("customOptions"))
                 .setId(call.<Integer>argument("adId"))
                 .setNativeAdOptions(call.<FlutterNativeAdOptions>argument("nativeAdOptions"))
-                .setFlutterAdLoader(new FlutterAdLoader(appContext))
+                .setFlutterAdLoader(new FlutterAdLoader(context))
                 .build();
         instanceManager.trackAd(nativeAd, call.<Integer>argument("adId"));
         nativeAd.load();
@@ -321,7 +384,7 @@ public class GoogleMobileAdsPlugin implements FlutterPlugin, ActivityAware, Meth
                 instanceManager,
                 call.<String>argument("adUnitId"),
                 call.<FlutterAdRequest>argument("request"),
-                new FlutterAdLoader(appContext));
+                new FlutterAdLoader(context));
         instanceManager.trackAd(interstitial, call.<Integer>argument("adId"));
         interstitial.load();
         result.success(null);
@@ -342,7 +405,7 @@ public class GoogleMobileAdsPlugin implements FlutterPlugin, ActivityAware, Meth
                   adUnitId,
                   request,
                   serverSideVerificationOptions,
-                  new FlutterAdLoader(appContext));
+                  new FlutterAdLoader(context));
         } else if (adManagerRequest != null) {
           rewardedAd =
               new FlutterRewardedAd(
@@ -351,7 +414,7 @@ public class GoogleMobileAdsPlugin implements FlutterPlugin, ActivityAware, Meth
                   adUnitId,
                   adManagerRequest,
                   serverSideVerificationOptions,
-                  new FlutterAdLoader(appContext));
+                  new FlutterAdLoader(context));
         } else {
           result.error("InvalidRequest", "A null or invalid ad request was provided.", null);
           break;
@@ -369,7 +432,7 @@ public class GoogleMobileAdsPlugin implements FlutterPlugin, ActivityAware, Meth
                 call.<String>argument("adUnitId"),
                 call.<List<FlutterAdSize>>argument("sizes"),
                 call.<FlutterAdManagerAdRequest>argument("request"),
-                getBannerAdCreator(appContext));
+                getBannerAdCreator(context));
         instanceManager.trackAd(adManagerBannerAd, call.<Integer>argument("adId"));
         adManagerBannerAd.load();
         result.success(null);
@@ -381,7 +444,7 @@ public class GoogleMobileAdsPlugin implements FlutterPlugin, ActivityAware, Meth
                 instanceManager,
                 call.<String>argument("adUnitId"),
                 call.<FlutterAdManagerAdRequest>argument("request"),
-                getBannerAdCreator(appContext));
+                getBannerAdCreator(context));
         instanceManager.trackAd(fluidAd, call.<Integer>argument("adId"));
         fluidAd.load();
         result.success(null);
@@ -393,7 +456,7 @@ public class GoogleMobileAdsPlugin implements FlutterPlugin, ActivityAware, Meth
                 requireNonNull(instanceManager),
                 requireNonNull(call.<String>argument("adUnitId")),
                 call.<FlutterAdManagerAdRequest>argument("request"),
-                new FlutterAdLoader(appContext));
+                new FlutterAdLoader(context));
         instanceManager.trackAd(
             adManagerInterstitialAd, requireNonNull(call.<Integer>argument("adId")));
         adManagerInterstitialAd.load();
@@ -407,8 +470,8 @@ public class GoogleMobileAdsPlugin implements FlutterPlugin, ActivityAware, Meth
                 requireNonNull(instanceManager),
                 requireNonNull(call.<String>argument("adUnitId")),
                 call.<FlutterAdRequest>argument("request"),
-                call.<FlutterAdManagerAdRequest>argument("adManagerAdRequest"),
-                new FlutterAdLoader(appContext));
+                call.<FlutterAdManagerAdRequest>argument("adManagerRequest"),
+                new FlutterAdLoader(context));
         instanceManager.trackAd(appOpenAd, call.<Integer>argument("adId"));
         appOpenAd.load();
         result.success(null);
@@ -428,7 +491,7 @@ public class GoogleMobileAdsPlugin implements FlutterPlugin, ActivityAware, Meth
       case "AdSize#getAnchoredAdaptiveBannerAdSize":
         final FlutterAdSize.AnchoredAdaptiveBannerAdSize size =
             new FlutterAdSize.AnchoredAdaptiveBannerAdSize(
-                appContext,
+                context,
                 new FlutterAdSize.AdSizeFactory(),
                 call.<String>argument("orientation"),
                 call.<Integer>argument("width"));
@@ -452,7 +515,7 @@ public class GoogleMobileAdsPlugin implements FlutterPlugin, ActivityAware, Meth
         result.success(null);
         break;
       case "MobileAds#disableMediationInitialization":
-        flutterMobileAds.disableMediationInitialization(appContext);
+        flutterMobileAds.disableMediationInitialization(context);
         result.success(null);
         break;
       case "MobileAds#getVersionString":
@@ -460,10 +523,18 @@ public class GoogleMobileAdsPlugin implements FlutterPlugin, ActivityAware, Meth
         break;
       case "getAdSize":
         FlutterAd ad = instanceManager.adForId(call.<Integer>argument("adId"));
-        if (ad instanceof FlutterBannerAd) {
+        if (ad == null) {
+          // This was called on a dart ad container that hasn't been loaded yet.
+          result.success(null);
+        } else if (ad instanceof FlutterBannerAd) {
           result.success(((FlutterBannerAd) ad).getAdSize());
         } else if (ad instanceof FlutterAdManagerBannerAd) {
           result.success(((FlutterAdManagerBannerAd) ad).getAdSize());
+        } else {
+          result.error(
+              Constants.ERROR_CODE_UNEXPECTED_AD_TYPE,
+              "Unexpected ad type for getAdSize: " + ad,
+              null);
         }
         break;
       default: