+++ packages/google_mobile_ads/ios/Tests/FLTGoogleMobileAdsReaderWriterTest.m
@@ -49,6 +49,20 @@
   XCTAssertEqualObjects(decodedSize.height, @(2));
 }
 
+- (void)testEncodeDecodeRequestConfiguration {
+  GADRequestConfiguration *requestConfiguration = [GADRequestConfiguration alloc];
+  requestConfiguration.maxAdContentRating = GADMaxAdContentRatingMatureAudience;
+  [GADMobileAds.sharedInstance.requestConfiguration tagForChildDirectedTreatment:YES];
+  [GADMobileAds.sharedInstance.requestConfiguration tagForUnderAgeOfConsent:NO];
+  NSArray<NSString *> *testDeviceIds = [[NSArray alloc] initWithObjects:@"test-device-id", nil];
+  requestConfiguration.testDeviceIdentifiers = testDeviceIds;
+  NSData *encodedMessage = [_messageCodec encode:requestConfiguration];
+
+  GADRequestConfiguration *decodedSize = [_messageCodec decode:encodedMessage];
+  XCTAssertEqualObjects(decodedSize.maxAdContentRating, GADMaxAdContentRatingMatureAudience);
+  XCTAssertEqualObjects(decodedSize.testDeviceIdentifiers, testDeviceIds);
+}
+
 - (void)testEncodeDecodeInlineAdaptiveBannerAdSize_currentOrientation {
   GADAdSize testAdSize = GADAdSizeFromCGSize(CGSizeMake(25, 10));
 
@@ -215,6 +229,8 @@
   NSArray<NSString *> *contentURLs = @[ @"url-1.com", @"url-2.com" ];
   request.neighboringContentURLs = contentURLs;
   request.location = [[FLTLocationParams alloc] initWithAccuracy:@1.5 longitude:@52 latitude:@123];
+  request.mediationExtrasIdentifier = @"identifier";
+  request.adMobExtras = @{@"key" : @"value"};
   NSData *encodedMessage = [_messageCodec encode:request];
 
   FLTAdRequest *decodedRequest = [_messageCodec decode:encodedMessage];
@@ -225,6 +241,8 @@
   XCTAssertEqualObjects(decodedRequest.location.accuracy, @1.5);
   XCTAssertEqualObjects(decodedRequest.location.longitude, @52);
   XCTAssertEqualObjects(decodedRequest.location.latitude, @123);
+  XCTAssertEqualObjects(decodedRequest.mediationExtrasIdentifier, @"identifier");
+  XCTAssertEqualObjects(decodedRequest.adMobExtras, @{@"key" : @"value"});
 }
 
 - (void)testEncodeDecodeGAMAdRequest {
@@ -238,6 +256,8 @@
   request.neighboringContentURLs = contentURLs;
   request.pubProvidedID = @"pub-id";
   request.location = [[FLTLocationParams alloc] initWithAccuracy:@1.5 longitude:@52 latitude:@123];
+  request.mediationExtrasIdentifier = @"identifier";
+  request.adMobExtras = @{@"key" : @"value"};
   NSData *encodedMessage = [_messageCodec encode:request];
 
   FLTGAMAdRequest *decodedRequest = [_messageCodec decode:encodedMessage];
@@ -252,6 +272,8 @@
   XCTAssertEqualObjects(decodedRequest.location.accuracy, @1.5);
   XCTAssertEqualObjects(decodedRequest.location.longitude, @52);
   XCTAssertEqualObjects(decodedRequest.location.latitude, @123);
+  XCTAssertEqualObjects(decodedRequest.mediationExtrasIdentifier, @"identifier");
+  XCTAssertEqualObjects(decodedRequest.adMobExtras, @{@"key" : @"value"});
 }
 
 - (void)testEncodeDecodeRewardItem {