+++ packages/google_mobile_ads/lib/src/ad_instance_manager.dart
@@ -611,6 +611,12 @@ class AdInstanceManager {
     );
   }
 
+  /// Gets the global [RequestConfiguration].
+  Future<RequestConfiguration> getRequestConfiguration() async {
+    return (await instanceManager.channel.invokeMethod<RequestConfiguration>(
+        'MobileAds#getRequestConfiguration'))!;
+  }
+
   /// Set the [RequestConfiguration] to apply for future ad requests.
   Future<void> updateRequestConfiguration(
       RequestConfiguration requestConfiguration) {
@@ -712,6 +718,7 @@ class AdMessageCodec extends StandardMessageCodec {
   static const int _valueVideoOptions = 145;
   static const int _valueInlineAdaptiveBannerAdSize = 146;
   static const int _valueLocationParams = 147;
+  static const int _valueRequestConfigurationParams = 148;
 
   @override
   void writeValue(WriteBuffer buffer, dynamic value) {
@@ -730,6 +737,8 @@ class AdMessageCodec extends StandardMessageCodec {
       }
       writeValue(buffer, value.publisherProvidedId);
       writeValue(buffer, value.location);
+      writeValue(buffer, value.mediationExtrasIdentifier);
+      writeValue(buffer, value.extras);
     } else if (value is AdRequest) {
       buffer.putUint8(_valueAdRequest);
       writeValue(buffer, value.keywords);
@@ -740,6 +749,8 @@ class AdMessageCodec extends StandardMessageCodec {
         writeValue(buffer, value.httpTimeoutMillis);
       }
       writeValue(buffer, value.location);
+      writeValue(buffer, value.mediationExtrasIdentifier);
+      writeValue(buffer, value.extras);
     } else if (value is RewardItem) {
       buffer.putUint8(_valueRewardItem);
       writeValue(buffer, value.amount);
@@ -795,6 +806,12 @@ class AdMessageCodec extends StandardMessageCodec {
       writeValue(buffer, value.clickToExpandRequested);
       writeValue(buffer, value.customControlsRequested);
       writeValue(buffer, value.startMuted);
+    } else if (value is RequestConfiguration) {
+      buffer.putUint8(_valueRequestConfigurationParams);
+      writeValue(buffer, value.maxAdContentRating);
+      writeValue(buffer, value.tagForChildDirectedTreatment);
+      writeValue(buffer, value.tagForUnderAgeOfConsent);
+      writeValue(buffer, value.testDeviceIds);
     } else if (value is LocationParams) {
       buffer.putUint8(_valueLocationParams);
       writeValue(buffer, value.accuracy);
@@ -872,6 +889,9 @@ class AdMessageCodec extends StandardMessageCodec {
               ? readValueOfType(buffer.getUint8(), buffer)
               : null,
           location: readValueOfType(buffer.getUint8(), buffer),
+          mediationExtrasIdentifier: readValueOfType(buffer.getUint8(), buffer),
+          extras: readValueOfType(buffer.getUint8(), buffer)
+              ?.cast<String, String>(),
         );
       case _valueRewardItem:
         return RewardItem(
@@ -921,6 +941,9 @@ class AdMessageCodec extends StandardMessageCodec {
               : null,
           publisherProvidedId: readValueOfType(buffer.getUint8(), buffer),
           location: readValueOfType(buffer.getUint8(), buffer),
+          mediationExtrasIdentifier: readValueOfType(buffer.getUint8(), buffer),
+          extras: readValueOfType(buffer.getUint8(), buffer)
+              ?.cast<String, String>(),
         );
       case _valueInitializationState:
         switch (readValueOfType(buffer.getUint8(), buffer)) {
@@ -970,6 +993,15 @@ class AdMessageCodec extends StandardMessageCodec {
           clickToExpandRequested: readValueOfType(buffer.getUint8(), buffer),
           customControlsRequested: readValueOfType(buffer.getUint8(), buffer),
           startMuted: readValueOfType(buffer.getUint8(), buffer),
+        );
+      case _valueRequestConfigurationParams:
+        return RequestConfiguration(
+          maxAdContentRating: readValueOfType(buffer.getUint8(), buffer),
+          tagForChildDirectedTreatment:
+              readValueOfType(buffer.getUint8(), buffer),
+          tagForUnderAgeOfConsent: readValueOfType(buffer.getUint8(), buffer),
+          testDeviceIds:
+              readValueOfType(buffer.getUint8(), buffer).cast<String>(),
         );
       case _valueLocationParams:
         return LocationParams(