+++ packages/google_mobile_ads/test/mobile_ads_test.dart
@@ -55,6 +55,15 @@ void main() {
             return null;
           case 'MobileAds#getVersionString':
             return Future<String>.value('Test-SDK-Version');
+          case 'MobileAds#updateRequestConfiguration':
+            return null;
+          case 'MobileAds#getRequestConfiguration':
+            return RequestConfiguration(
+              maxAdContentRating: MaxAdContentRating.ma,
+              tagForChildDirectedTreatment: TagForChildDirectedTreatment.yes,
+              tagForUnderAgeOfConsent: TagForUnderAgeOfConsent.no,
+              testDeviceIds: <String>['test-device-id'],
+            );
           case 'AdSize#getAnchoredAdaptiveBannerAdSize':
             return null;
           default:
@@ -309,6 +318,78 @@ void main() {
             arguments: {'orientation': 'landscape', 'width': 34}),
         isMethodCall('AdSize#getAnchoredAdaptiveBannerAdSize',
             arguments: {'width': 45})
+      ]);
+    });
+
+    test('encode/decode $MobileAds.getRequestConfiguration', () async {
+      RequestConfiguration requestConfig =
+          await MobileAds.instance.getRequestConfiguration();
+
+      debugDefaultTargetPlatformOverride = TargetPlatform.iOS;
+      ByteData byteData = codec.encodeMessage(requestConfig)!;
+      RequestConfiguration result = codec.decodeMessage(byteData);
+
+      expect(result.maxAdContentRating, MaxAdContentRating.ma);
+      expect(result.tagForChildDirectedTreatment,
+          TagForChildDirectedTreatment.yes);
+      expect(result.tagForUnderAgeOfConsent, TagForUnderAgeOfConsent.no);
+      expect(result.testDeviceIds, ['test-device-id']);
+
+      debugDefaultTargetPlatformOverride = TargetPlatform.android;
+      byteData = codec.encodeMessage(requestConfig)!;
+      result = codec.decodeMessage(byteData);
+
+      expect(result.maxAdContentRating, MaxAdContentRating.ma);
+      expect(result.tagForChildDirectedTreatment,
+          TagForChildDirectedTreatment.yes);
+      expect(result.tagForUnderAgeOfConsent, TagForUnderAgeOfConsent.no);
+      expect(result.testDeviceIds, ['test-device-id']);
+    });
+
+    test('$MobileAds.RequestConfiguration', () async {
+      RequestConfiguration requestConfiguration = RequestConfiguration(
+        maxAdContentRating: MaxAdContentRating.ma,
+        tagForChildDirectedTreatment: TagForChildDirectedTreatment.yes,
+        tagForUnderAgeOfConsent: TagForUnderAgeOfConsent.no,
+        testDeviceIds: <String>['test-device-id'],
+      );
+
+      await MobileAds.instance.updateRequestConfiguration(requestConfiguration);
+
+      debugDefaultTargetPlatformOverride = TargetPlatform.iOS;
+      RequestConfiguration result =
+          await MobileAds.instance.getRequestConfiguration();
+      expect(result.maxAdContentRating, MaxAdContentRating.ma);
+      expect(result.tagForChildDirectedTreatment,
+          TagForChildDirectedTreatment.yes);
+      expect(result.tagForUnderAgeOfConsent, TagForUnderAgeOfConsent.no);
+      expect(result.testDeviceIds, <String>['test-device-id']);
+      expect(log, <Matcher>[
+        isMethodCall('MobileAds#updateRequestConfiguration', arguments: {
+          'maxAdContentRating': MaxAdContentRating.ma,
+          'tagForChildDirectedTreatment': TagForChildDirectedTreatment.yes,
+          'testDeviceIds': <String>['test-device-id'],
+          'tagForUnderAgeOfConsent': TagForUnderAgeOfConsent.no,
+        }),
+        isMethodCall('MobileAds#getRequestConfiguration', arguments: null)
+      ]);
+
+      debugDefaultTargetPlatformOverride = TargetPlatform.android;
+      result = await MobileAds.instance.getRequestConfiguration();
+      expect(result.maxAdContentRating, MaxAdContentRating.ma);
+      expect(result.tagForChildDirectedTreatment,
+          TagForChildDirectedTreatment.yes);
+      expect(result.tagForUnderAgeOfConsent, TagForUnderAgeOfConsent.no);
+      expect(result.testDeviceIds, <String>['test-device-id']);
+      expect(log, <Matcher>[
+        isMethodCall('MobileAds#updateRequestConfiguration', arguments: {
+          'maxAdContentRating': MaxAdContentRating.ma,
+          'tagForChildDirectedTreatment': TagForChildDirectedTreatment.yes,
+          'testDeviceIds': <String>['test-device-id'],
+          'tagForUnderAgeOfConsent': TagForUnderAgeOfConsent.no,
+        }),
+        isMethodCall('MobileAds#getRequestConfiguration', arguments: null),
+        isMethodCall('MobileAds#getRequestConfiguration', arguments: null)
       ]);
     });
   });