+++ packages/google_mobile_ads/CHANGELOG.md
@@ -1,3 +1,45 @@
+## 1.0.1
+
+* Fix for [Issue 449](https://github.com/googleads/googleads-mobile-flutter/issues/449).
+  In `LocationParams`, time is now treated as an optional parameter on Android.
+* Fix for [Issue 447](https://github.com/googleads/googleads-mobile-flutter/issues/447),
+  which affected mediation networks that require an Activity to be initialized on Android.
+
+## 1.0.0
+
+* Mediation is now supported in beta.
+  * There are new APIs to support passing network extras to mediation adapters:
+    * [MediationNetworkExtrasProvider](https://github.com/googleads/googleads-mobile-flutter/blob/master/packages/google_mobile_ads/android/src/main/java/io/flutter/plugins/googlemobileads/MediationNetworkExtrasProvider.java)
+      on Android and [FLTMediationNetworkExtrasProvider](https://github.com/googleads/googleads-mobile-flutter/blob/master/packages/google_mobile_ads/ios/Classes/FLTConstants.h) on iOS
+  * See the mediation example app [README](https://github.com/googleads/googleads-mobile-flutter/blob/master/packages/mediation_example/README.md)
+    for more details on how to use these APIs.
+
+* Fix for Android 12 issue [#330](https://github.com/googleads/googleads-mobile-flutter/issues/330)
+  * This will break compilation on android if you do not already set `compileSdkVersion` to `31`, or override the WorkManager dependency to < 2.7.0:
+      ```
+      dependencies {
+          implementation('androidx.work:work-runtime') {
+              version {
+                  strictly '2.6.0'
+              }
+          }
+      }
+      ```
+* Fixes issue [#404](https://github.com/googleads/googleads-mobile-flutter/issues/404)
+  * Adds a new dart class, `AppStateEventNotifier`. You should subscribe to `AppStateEventNotifier.appStateStream`
+    instead of using `WidgetsBindingObserver` to listen to app foreground/background events.
+  * See the app open [example app](https://github.com/googleads/googleads-mobile-flutter/tree/master/packages/app_open_example) for a reference
+    on how to use the new API.
+    
+* Adds a new parameter `extras` to `AdRequest` and `AdManagerAdRequest`.
+  * This can be used to pass additional signals to the AdMob adapter, such as
+    [CCPA](https://developers.google.com/admob/android/ccpa) signals.
+  * For example, to notify Google that [RDP](https://developers.google.com/admob/android/ccpa#rdp_signal)
+    should be enabled when constructing an ad request:
+    ```dart
+      AdRequest request = AdRequest(extras: {'rdp': '1'});
+    ```
+    
 ## 0.13.6
 
 * Partial fix for [#265](https://github.com/googleads/googleads-mobile-flutter/issues/265).
@@ -5,13 +47,16 @@
     but it only works the first time the engine is attached to an activity.
   * Support for reusing the engine in another activity after the first one is destroyed is blocked 
     by this Flutter issue which affects all platform views: https://github.com/flutter/flutter/issues/88880.
+* Adds support for getRequestConfiguration API
+  * [Android API reference](https://developers.google.com/android/reference/com/google/android/gms/ads/MobileAds#public-static-requestconfiguration-getrequestconfiguration)
+  * [iOS API reference](https://developers.google.com/admob/ios/api/reference/Classes/GADMobileAds#requestconfiguration)
 * Adds support for Fluid Ad Size (Ad Manager only)
   * Fluid ads dynamically adjust their height based on their width. To help display them we've added a new
     ad container, `FluidAdManagerBannerAd`, and a new widget `FluidAdWidget`.
-    You can see the [fluid_example.dart](https://github.com/googleads/googleads-mobile-flutter/blob/master/packages/google_mobile_ads/example/lib/fluid_example.dart) for a reference of how to load and display a fluid ad.
-  * Android - https://developers.google.com/ad-manager/mobile-ads-sdk/android/native/styles#fluid_size
-  * iOS - https://developers.google.com/ad-manager/mobile-ads-sdk/ios/native/native-styles#fluid_size
-* Adds `AdSize. getCurrentOrientationAnchoredAdaptiveBannerAdSize()` to support getting an `AnchoredAdaptiveBannerAdSize` in the current orientation.
+  * You can see the [fluid_example.dart](https://github.com/googleads/googleads-mobile-flutter/blob/master/packages/google_mobile_ads/example/lib/fluid_example.dart) for a reference of how to load and display a fluid ad.
+  * [Android API reference](https://developers.google.com/ad-manager/mobile-ads-sdk/android/native/styles#fluid_size)
+  * [iOS API reference](https://developers.google.com/ad-manager/mobile-ads-sdk/ios/native/native-styles#fluid_size)
+* Adds `AdSize.getCurrentOrientationAnchoredAdaptiveBannerAdSize()` to support getting an `AnchoredAdaptiveBannerAdSize` in the current orientation.
   * Previously the user had to specify an orientation (portrait / landscape) to create an AnchoredAdaptiveBannerAdSize. It has been made optional with this version. SDK will determine the current orientation of the device and return an appropriate AdSize.
   * More information on anchored adaptive banners can be found here:
     * [Admob android](https://developers.google.com/admob/android/banner/anchored-adaptive)